Fix cross login issue

+++ base/views.py
@@ -1,7 +1,11 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
 
@@ -14,17 +18,58 @@ from .forms import RoomForm
 
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
 
+    # Get username and password
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
+        # check if the user exists
         try:
             user = User.objects.get(username=username)
         except:
             messages.error(request, 'User does not exist!')
 
-    context = {}
+        # Authentication
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            # logins the user
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or Password is not correct!')
+
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            if form.is_valid():
+                user = form.save(commit=False)
+                user.username = user.username.lower()
+                user.save()
+                # simultaneously login the user
+                login(request, user)
+                return redirect('home')
+        else:
+            messages.error(request, 'An Error occurred during Registraion')
+
+
+    context = {'form':form}
     return render(request, 'base/login_register.html', context)
 
 
@@ -54,6 +99,7 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
 
@@ -67,10 +113,16 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     # `instance` is so that the form we get is pre filled
     form = RoomForm(instance=room)
+
+    # To prevent someone else from accessing other's room
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
+
 
     if request.method == 'POST':
         # `instance` here will replace the pre filled value
@@ -83,8 +135,14 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    # To prevent someone else from accessing other's room
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
